--- nihon/src/app/admin/layout.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const metadata = {
-  title: "Admin - Login",
-};
-
-export default function AdminLayout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
